refactor(runtime): hoist RuntimeOptions interfaces to runtime-types.ts (W26-C step 1)
Moves MCPServerConfig (93 LOC), RuntimeOptions (641 LOC), and LightRuntimeOptions
(36 LOC) verbatim into runtime-types.ts. runtime.ts re-exports all three names
for backward compatibility with builder.ts and external consumers.

runtime.ts: 1997 -> 1218 LOC (-779 / -39%). Below the 1500 threshold from #76.
Runtime 811/0/1, replay 24/0, build 38/38.

Partial: #76
